commit to work on chromebook

+++ src/main.rs
@@ -1,7 +1,7 @@
 use clap::Parser;
 use rand::{Rng, SeedableRng};
 use rust_pkg_gen::resources::TemplateAssets;
-use std::{fs, path::PathBuf};
+use std::{fs, path::PathBuf, process::Stdio};
 
 fn gen_char() -> u8 {
     rand::rngs::StdRng::from_entropy().gen_range(65..90)
@@ -31,8 +31,30 @@ struct Cli {
         help = "The equivalent to --yes for overwriting files"
     )]
     overwrite: bool,
-    #[arg(default_value = "rust-config.toml", help = "")]
+    #[arg(
+        default_value = "rust-config.toml",
+        help = "The path to the configuration file"
+    )]
     path: PathBuf,
+    #[arg(
+        short = 'q',
+        long = "quiet",
+        default_value_t = false,
+        help = "Doesn't display any unnecessary text(still shows confirmation prompts; to remove, use --overwrite"
+    )]
+    quiet: bool,
+    #[arg(
+        long = "silent",
+        default_value_t = false,
+        help = "Equivalent to -y -q --overwrite"
+    )]
+    silent: bool,
+    #[arg(
+        long = "save-temp",
+        default_value_t = false,
+        help = "Saves all temporary files"
+    )]
+    save_temp: bool,
 }
 
 fn generate_crates(
@@ -47,7 +69,13 @@ fn generate_crates(
 }
 
 fn main() {
-    let args = Cli::parse();
+    let mut args = Cli::parse();
+
+    if args.silent {
+        args.quiet = true;
+        args.yes = true;
+        args.overwrite = true;
+    }
 
     let path = &PathBuf::from(args.path);
 
@@ -62,17 +90,24 @@ fn main() {
     let dir = if args.temp_dir.is_some() {
         if args.temp_dir.clone().unwrap().exists() && !args.overwrite {
             let confirmation = dialoguer::Confirm::new()
-                .with_prompt("Provided temporary directory already exists, overwrite?")
+                .with_prompt("Temporary directory already exists, overwrite?")
                 .default(false)
                 .interact()
                 .unwrap();
             if !confirmation {
-                println!("Aborting.");
+                if !args.quiet {
+                    println!("Aborting.");
+                }
                 std::process::exit(1)
             } else {
-                println!("Overwriting.");
+                if !args.quiet {
+                    println!("Overwriting.");
+                }
                 std::fs::remove_dir_all(args.temp_dir.clone().unwrap()).unwrap();
             }
+        }
+        if args.overwrite {
+            std::fs::remove_dir_all(args.temp_dir.clone().unwrap()).unwrap();
         }
         args.temp_dir.unwrap()
     } else {
@@ -115,15 +150,21 @@ fn main() {
                 .unwrap();
             }
 
-            std::process::Command::new("bash")
-                .arg(
-                    dir.join(PathBuf::from("crates"))
-                        .join(PathBuf::from("build.sh")),
-                )
-                .spawn()
-                .unwrap()
-                .wait()
-                .unwrap();
+            let mut build_c = std::process::Command::new("bash");
+            let mut build = build_c.arg(
+                dir.join(PathBuf::from("crates"))
+                    .join(PathBuf::from("build.sh")),
+            );
+            if args.quiet {
+                build = build
+                    .stdout(Stdio::null())
+                    .stderr(Stdio::null())
+                    .stdin(Stdio::null());
+            }
+            if args.save_temp {
+                build = build.arg("save")
+            }
+            build.spawn().unwrap().wait().unwrap();
 
             rust_pkg_gen::copied::download_all(
                 vec![&toolchain.channel],
@@ -132,10 +173,13 @@ fn main() {
                 toolchain.targets.iter().map(|s| &**s).collect(),
                 dir.join("toolchain").to_str().unwrap(),
                 toolchain.components.iter().map(|s| &**s).collect(),
-                toolchain.platforms.iter().map(|s| &**s).collect()
+                toolchain.platforms.iter().map(|s| &**s).collect(),
+                args.quiet,
             );
 
-            fs::remove_dir_all(dir.join("tmp").to_str().unwrap()).unwrap();
+            if !args.save_temp {
+                fs::remove_dir_all(dir.join("tmp").to_str().unwrap()).unwrap();
+            }
         }
     }
 }
